Add 1.4 large VDM.

+++ idekm/src/pci_ide_km_requester/pci_ide_km_req_key_prog.rs
@@ -54,7 +54,7 @@ impl IdekmReqContext {
         }
         .encode(&mut writer)
         .map_err(|_| SPDM_STATUS_BUFFER_FULL)?
-            as u16;
+            as u32;
 
         let vendor_defined_rsp_payload_struct = spdm_requester
             .send_spdm_vendor_defined_request(
