+++ idekm/src/pci_ide_km_responder/pci_ide_km_rsp_key_set_go.rs
@@ -103,7 +103,7 @@ pub(crate) fn pci_ide_km_rsp_key_set_go(
     if cnt > u16::MAX as usize {
         Err(SPDM_STATUS_INVALID_STATE_LOCAL)
     } else {
-        vendor_defined_rsp_payload_struct.rsp_length = cnt as u16;
+        vendor_defined_rsp_payload_struct.rsp_length = cnt as u32;
         Ok(vendor_defined_rsp_payload_struct)
     }
 }
