+++ tdisp/src/pci_tdisp_responder/pci_tdisp_rsp_device_interface_state.rs
@@ -102,7 +102,7 @@ pub(crate) fn pci_tdisp_rsp_interface_state(
             &[],
             &mut vendor_defined_rsp_payload_struct.vendor_defined_rsp_payload,
         )?;
-        vendor_defined_rsp_payload_struct.rsp_length = len as u16;
+        vendor_defined_rsp_payload_struct.rsp_length = len as u32;
         return Ok(vendor_defined_rsp_payload_struct);
     }
 
@@ -126,7 +126,7 @@ pub(crate) fn pci_tdisp_rsp_interface_state(
     if cnt > u16::MAX as usize {
         Err(SPDM_STATUS_INVALID_STATE_LOCAL)
     } else {
-        vendor_defined_rsp_payload_struct.rsp_length = cnt as u16;
+        vendor_defined_rsp_payload_struct.rsp_length = cnt as u32;
         Ok(vendor_defined_rsp_payload_struct)
     }
 }
