+++ tdisp/src/pci_tdisp_responder/pci_tdisp_rsp_stop_interface_request.rs
@@ -93,7 +93,7 @@ pub(crate) fn pci_tdisp_rsp_stop_interface(
             &[],
             &mut vendor_defined_rsp_payload_struct.vendor_defined_rsp_payload,
         )?;
-        vendor_defined_rsp_payload_struct.rsp_length = len as u16;
+        vendor_defined_rsp_payload_struct.rsp_length = len as u32;
         return Ok(vendor_defined_rsp_payload_struct);
     }
 
@@ -116,7 +116,7 @@ pub(crate) fn pci_tdisp_rsp_stop_interface(
     if cnt > u16::MAX as usize {
         Err(SPDM_STATUS_INVALID_STATE_LOCAL)
     } else {
-        vendor_defined_rsp_payload_struct.rsp_length = cnt as u16;
+        vendor_defined_rsp_payload_struct.rsp_length = cnt as u32;
         Ok(vendor_defined_rsp_payload_struct)
     }
 }
