+++ tdisp/src/pci_tdisp_responder/pci_tdisp_rsp_tdisp_version.rs
@@ -113,7 +113,7 @@ pub(crate) fn pci_tdisp_rsp_version(
     if cnt > u16::MAX as usize {
         Err(SPDM_STATUS_INVALID_STATE_LOCAL)
     } else {
-        vendor_defined_rsp_payload_struct.rsp_length = cnt as u16;
+        vendor_defined_rsp_payload_struct.rsp_length = cnt as u32;
         Ok(vendor_defined_rsp_payload_struct)
     }
 }
